// Desktop/cs5500-gui-arch-huangweili97-main/src/types/types.tsx
export interface AnswerPropsInterface
 {
    aid: string;
    text: string;
    ansBy: string;
    ansDate: Date;
    calculateTimeElapsed: () => string;
}

/**
 * An interface representing the inupt parameters 
 * the constructor of the Answer class expects
 */
export interface AnswerParamType
 {
    aid: string;
    text: string;
    ansBy: string;
    ansDate: Date;
}